feat(admin): ship Slice 7 quotas and administration
Implements: AC-3–AC-6, AD-1–AD-6, AD-8, AD-9 (TRD §13, §14, §17 row 7).

Verified: real price-weighted model accounting, atomic quota reservation and actual settlement, Deep budget enforcement, versioned settings rollback with audit logging, composer quota UX, and admin CRUD.

Confirmed: no TODO(slice-7) markers remain.

+++ apps/api/schemas/events.py
@@ -66,8 +66,7 @@ class Metrics(RunEvent):
     latency_ms: dict[str, int] = {}
     tokens_in: int = 0
     tokens_out: int = 0
-    # Token counts until the ledger lands (slice 7, TRD §14).
-    credits: int = 0
+    credits: float = 0
     context_used: int = 0
     context_window: int = 0
     faithfulness: float | None = None
